Creating a node::Log module
create a new file in src/raft/node/log.rs

this file is hold the logic for nodes/peers log structure
